refactor event interfaces
