--- frontend/interfaces/filterEvents.tsx
@@ -1,15 +1,0 @@
-export interface filterEvents {
-  groups: number[],
-  eventTypes: number[],
-  gameSystems: number[],
-  ageRequirements: number[],
-  experienceRequirements: number[],
-  startDates: number[],
-  startTimes: number[],
-  endDates: number[],
-  endTimes: number[],
-  ifTournament: number[],
-  costs: number[],
-  locations: number[],
-  ticketsAvailable: number[]
-}
