+++ frontend/pages/index.tsx
@@ -10,16 +10,6 @@ import TimeRange from "@/components/TimeRange";
 
 export const { eventData, filters } = getData();
 
-export interface Choice {
-  eventId: number;
-  endDate: string;
-  endTime: string;
-  startDate: string;
-  startTime: string;
-  duration: number;
-  cost: number;
-}
-
 export default function Home() {
   const filterList = filters;
 
@@ -48,7 +38,7 @@ export default function Home() {
   const [hideSoldOut, setHideSoldOut] = useState(false);
 
   // Choices
-  const [choices, setChoices] = useState<Choice[]>([])
+  const [choices, setChoices] = useState<number[]>([])
 
   return (
     <main>
